requirements.txt
Pillow==9.4.0
pyinstaller==5.9.0
pypng==0.20220715.0
qrcode==7.4.2
xlwt==1.3.0
xlrd==2.0.1